Removed the pseudocode
At this point, I don't see how the pseudocode is going to work out for this project. Since each of us created a separate branch in order to implement each of the parts for the game, the pseudocode ended up being irrelavant to the repository. Therefore, the pseudocode for 10000 dice is no longer necessary for the project.